requirements.txt
asgiref==3.8.1
cffi==1.17.1
cryptography==43.0.3
Django==5.1.3
djangorestframework==3.15.2
djangorestframework-simplejwt==5.3.1
pillow==11.0.0
pycparser==2.22
PyJWT==2.10.0
sqlparse==0.5.2
typing_extensions==4.12.2
tzdata==2024.2
